Refactor and simplify thread feed cards
