+++ src/views/dashboard/style.js
@@ -1,6 +1,5 @@
 // @flow
 import styled, { css, keyframes } from 'styled-components';
-import Link from 'src/components/link';
 import {
   zIndex,
   Truncate,
@@ -360,338 +359,6 @@ export const ComposeIconContainer = styled.div`
 
   div {
     margin-right: 8px;
-  }
-`;
-
-export const InboxThreadItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  max-width: 100%;
-  min-width: 0;
-  overflow-x: hidden;
-  border-bottom: 1px solid
-    ${props => (props.active ? props.theme.brand.alt : props.theme.bg.border)};
-  background: ${props =>
-    props.active ? props.theme.brand.alt : props.theme.bg.default};
-  position: relative;
-
-  &:hover {
-    background: ${props =>
-      props.active ? props.theme.brand.alt : props.theme.bg.wash};
-  }
-
-  &:last-of-type {
-    border-bottom: 1px solid
-      ${props => (props.active ? props.theme.brand.alt : props.theme.bg.border)};
-  }
-`;
-
-export const InboxLinkWrapper = styled(Link)`
-  position: absolute;
-  display: inline-block;
-  height: 100%;
-  width: 100%;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: ${zIndex.card};
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-export const InboxClickWrapper = styled.span`
-  position: absolute;
-  display: inline-block;
-  height: 100%;
-  width: 100%;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: ${zIndex.card};
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-export const InboxThreadContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-self: stretch;
-  position: relative;
-  z-index: ${zIndex.card + 1};
-  align-items: flex-start;
-  pointer-events: none;
-`;
-
-export const ThreadTitle = styled.h3`
-  font-size: 16px;
-  font-weight: 500;
-  color: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.text.default};
-  margin: 12px 16px 0;
-  max-width: 100%;
-  line-height: 1.4;
-`;
-
-export const AttachmentsContainer = styled.div`
-  margin: 4px 16px;
-  max-width: calc(100% - 40px);
-
-  a {
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.alt};
-  }
-
-  a:hover {
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.default};
-  }
-`;
-
-export const ThreadMeta = styled.div`
-  align-self: stretch;
-  display: flex;
-  margin: 8px 16px;
-  justify-content: flex-start;
-  align-items: center;
-`;
-
-export const ThreadAuthorWrapper = styled.div`
-  margin-left: 8px;
-  ${Tooltip};
-`;
-
-export const ThreadActivityWrapper = styled.div`
-  display: flex;
-  flex: none;
-  align-items: center;
-
-  .icon {
-    pointer-events: auto;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
-`;
-
-export const ThreadStatusWrapper = styled(ThreadActivityWrapper)`
-  flex: auto;
-  justify-content: flex-start;
-  color: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.text.alt};
-
-  .icon {
-    pointer-events: auto;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
-`;
-
-export const CountWrapper = styled.div`
-  display: flex;
-  flex: none;
-  font-size: 14px;
-  color: ${props =>
-    props.new
-      ? props.active ? props.theme.text.reverse : props.theme.warn.alt
-      : props.active ? props.theme.text.reverse : props.theme.text.alt};
-  font-weight: ${props => (props.new ? 600 : 400)};
-  align-items: center;
-  margin-right: 8px;
-
-  > span {
-    margin-left: 4px;
-  }
-
-  a {
-    font-weight: 600;
-  }
-
-  a:hover {
-    color: ${props => props.theme.text.default};
-  }
-`;
-
-const StatusPill = styled(CountWrapper)`
-  color: ${props =>
-    props.active ? props.theme.brand.alt : props.theme.text.reverse};
-  background: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.brand.alt};
-  border-radius: 20px;
-  padding: 4px 12px;
-  font-size: 11px;
-  text-transform: uppercase;
-  font-weight: 700;
-  display: flex;
-  align-items: center;
-`;
-
-export const NewThreadPill = styled(StatusPill)`
-  color: ${props =>
-    props.active ? props.theme.brand.alt : props.theme.text.reverse};
-  background: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.success.alt};
-`;
-
-export const NewMessagePill = styled(CountWrapper)`
-  color: ${props =>
-    props.active ? props.theme.brand.alt : props.theme.warn.alt};
-`;
-
-export const LockedTextPill = styled(StatusPill)`
-  color: ${props =>
-    props.active ? props.theme.brand.alt : props.theme.text.alt};
-  background: ${props =>
-    props.active ? props.theme.brand.wash : props.theme.bg.border};
-`;
-
-export const MetaCommunityName = styled(Link)`
-  padding-right: 8px;
-  pointer-events: auto;
-  white-space: nowrap;
-`;
-
-export const CommunityInfoContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px 16px 0;
-  width: 100%;
-
-  a {
-    font-size: 14px;
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.alt};
-  }
-
-  a.pill {
-    font-size: 12px;
-    box-shadow: 0 0 0 1px
-      ${props =>
-        props.active ? props.theme.text.reverse : props.theme.bg.border};
-    background: ${props =>
-      props.active ? 'transparent' : props.theme.bg.wash};
-    font-weight: ${props => (props.active ? '500' : '400')};
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.alt};
-  }
-
-  a:hover {
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.default};
-  }
-
-  a.pill:hover {
-    background: ${props =>
-      props.active ? 'rgba(255,255,255,0.1)' : props.theme.bg.wash};
-  }
-`;
-
-export const LastActiveTimestamp = styled.p`
-  font-size: 14px;
-  font-weight: 400;
-  color: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.text.alt};
-  line-height: 1.28;
-  pointer-events: auto;
-`;
-
-export const AvatarLink = styled(Link)`
-  display: inline-block;
-  width: 20px;
-  height: 20px;
-  margin-right: 8px;
-  border-radius: 4px;
-  pointer-events: auto;
-  overflow: hidden;
-`;
-
-export const CommunityAvatar = styled.img`
-  width: 100%;
-  height: 100%;
-  pointer-events: auto;
-`;
-
-export const CommunityLink = styled(Link)`
-  font-size: 14px;
-  font-weight: 400;
-  color: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.text.alt};
-  line-height: 1.28;
-  pointer-events: auto;
-
-  &:hover {
-    color: ${props => props.theme.text.default};
-  }
-`;
-
-export const PillLink = styled(Link)`
-  display: inline-block;
-  display: flex;
-  align-items: center;
-  border-radius: 4px;
-  overflow: hidden;
-  padding: 4px 8px;
-  margin-right: 8px;
-  font-size: 12px;
-  max-height: 24px;
-  line-height: 1;
-  pointer-events: auto;
-`;
-
-export const PillLinkPinned = styled.div`
-  background: ${props => props.theme.special.wash};
-  border: 1px solid ${props => props.theme.special.border};
-  color: ${props => props.theme.special.dark};
-  display: flex;
-  height: 20px;
-  border-radius: 4px;
-  overflow: hidden;
-  padding: 0 8px;
-  margin-right: 8px;
-  font-size: 12px;
-  max-height: 24px;
-  line-height: 1.5;
-  .icon {
-    top: -1px;
-  }
-`;
-
-export const PillLabel = styled.span`
-  overflow: hidden;
-  max-width: 128px;
-  text-overflow: ellipsis;
-  display: inline-block;
-  white-space: nowrap;
-`;
-
-export const MiniLinkPreview = styled.a`
-  display: flex;
-  flex: auto;
-  align-items: center;
-  font-size: 14px;
-  color: ${props =>
-    props.active ? props.theme.text.reverse : props.theme.text.alt};
-  font-weight: ${props => (props.active ? '500' : '400')};
-  margin-bottom: 4px;
-  pointer-events: auto;
-  min-width: 0;
-
-  .icon {
-    margin-right: 4px;
-  }
-
-  &:hover {
-    color: ${props =>
-      props.active ? props.theme.text.reverse : props.theme.text.default};
   }
 `;
